Split AccountDataStore and TagStore
